Доработка общего кода
