+++ RegisterWindow.xaml.cs
@@ -29,5 +29,11 @@ namespace Cats_Program
             SignUpWindow signUpWindow = new SignUpWindow();
             signUpWindow.Show();
         }
+
+        private void btLogin_Click(object sender, RoutedEventArgs e)
+        {
+            Reg reg = new Reg();
+            reg.Show();
+        }
     }
 }
